cleaned up ansys launcher for CORBA for windows
