+++ tests/test_mapdl.py
@@ -23,6 +23,25 @@ if not HAS_ANSYS:
 skip_no_xserver = pytest.mark.skipif(not system_supports_plotting(),
                                      reason="Requires active X Server")
 
+@pytest.fixture(scope="module", params=['console', 'corba'])
+def mapdl(request):
+    os.environ['I_MPI_SHM_LMT'] = 'shm'  # necessary on ubuntu
+
+    mode = request.param
+    rver = '202'
+    mapdl = pyansys.launch_mapdl(get_ansys_bin(rver), override=True, mode=mode)
+    mapdl._show_matplotlib_figures = False
+    return mapdl
+
+
+@pytest.fixture(scope='function')
+def cleared(mapdl):
+    mapdl.finish()
+    mapdl.clear('NOSTART')  # *MUST* be NOSTART.  With START fails after 20 calls...
+    mapdl.prep7()
+    yield
+
+
 
 @pytest.fixture(scope='function')
 def make_block(mapdl, cleared):
